+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/AddressApiDtoMapper.java
@@ -1,16 +1,17 @@
-package club.klabis.adapters.api.mappers;
+package club.klabis.adapters.api.mappers.members;
 
-import club.klabis.api.dto.SexApiDto;
+import club.klabis.api.dto.AddressApiDto;
 import club.klabis.common.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.Sex;
+import club.klabis.domain.members.Address;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface SexApiDtoMapper extends Converter<Sex, SexApiDto> {
+interface AddressApiDtoMapper extends Converter<Address, AddressApiDto> {
     @DelegatingConverter
     @InheritInverseConfiguration
-    Sex fromApiDto(SexApiDto apiDto);
+    Address fromApiDto(AddressApiDto apiDto);
+
 }
