+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/IdentityCardApiDtoMapper.java
@@ -1,4 +1,4 @@
-package club.klabis.adapters.api.mappers;
+package club.klabis.adapters.api.mappers.members;
 
 import club.klabis.api.dto.IdentityCardApiDto;
 import club.klabis.common.DomainToDtoMapperConfiguration;
