+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/RefereeLicenceApiDtoMapper.java
@@ -1,4 +1,4 @@
-package club.klabis.adapters.api.mappers;
+package club.klabis.adapters.api.mappers.members;
 
 
 import club.klabis.api.dto.RefereeLicenceApiDto;
