+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/SexApiDtoMapper.java
@@ -1,17 +1,16 @@
-package club.klabis.adapters.api.mappers;
+package club.klabis.adapters.api.mappers.members;
 
-import club.klabis.api.dto.AddressApiDto;
+import club.klabis.api.dto.SexApiDto;
 import club.klabis.common.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.Address;
+import club.klabis.domain.members.Sex;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface AddressApiDtoMapper extends Converter<Address, AddressApiDto> {
+interface SexApiDtoMapper extends Converter<Sex, SexApiDto> {
     @DelegatingConverter
     @InheritInverseConfiguration
-    Address fromApiDto(AddressApiDto apiDto);
-
+    Sex fromApiDto(SexApiDto apiDto);
 }
