+++ backend/src/main/java/club/klabis/adapters/api/mappers/members/TrainerLicenceApiDtoMapper.java
@@ -1,8 +1,9 @@
-package club.klabis.adapters.api.mappers;
+package club.klabis.adapters.api.mappers.members;
 
-import club.klabis.api.dto.LegalGuardianApiDto;
+
+import club.klabis.api.dto.TrainerLicenceApiDto;
 import club.klabis.common.DomainToDtoMapperConfiguration;
-import club.klabis.domain.members.LegalGuardian;
+import club.klabis.domain.members.TrainerLicence;
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -10,14 +11,12 @@ import org.mapstruct.extensions.spring.DelegatingConverter;
 import org.springframework.core.convert.converter.Converter;
 
 @Mapper(config = DomainToDtoMapperConfiguration.class)
-interface LegalGuardianApiDtoMapper extends Converter<LegalGuardian, LegalGuardianApiDto> {
-
+interface TrainerLicenceApiDtoMapper extends Converter<TrainerLicence, TrainerLicenceApiDto> {
     @Override
-    @Mapping(source = "contacts", target = "contact")
-    LegalGuardianApiDto convert(LegalGuardian source);
+    @Mapping(target = "licence", source = "type")
+    TrainerLicenceApiDto convert(TrainerLicence source);
 
     @DelegatingConverter
     @InheritInverseConfiguration
-    LegalGuardian fromApiDto(LegalGuardianApiDto apiDto);
-
+    TrainerLicence fromApiDto(TrainerLicenceApiDto apiDto);
 }
